NEXT-13481 - Use indexed columns in stock updater to reduce the search space

+++ src/Elasticsearch/Profiler/DataCollector.php
@@ -49,7 +49,7 @@ class DataCollector extends BaseDataCollector
             $time += $calledRequest['time'];
         }
 
-        return $time * 1000;
+        return (int) ($time * 1000);
     }
 
     public function getRequestAmount(): int
